Uploading current state of the game.
Rendering of the canvas, missile bases and base guns is setup. 
Class framework for missiles and explosions in place but not fully implemented. 
Missiles are created on mouseclick but not yet rendered or updated, only added to missile array to be updated in future. 
Active missile base currently determined by which sector of the screen the mouse is in.
Mouse cursor implemented and to be used in place of standard mouse cursor.
Game has rudimentary pause function with overlay to indicate pause is in effect.
As of yet no score function is implemented and will likely remain like this until full game loop implemented.
Still to create and develop COM ability to spawn missiles aimed at player structures as well as "bonus" entities such as planes and UFOs.
In addition to the above, once enemy missiles handled, will need to also handle ability of enemy missiles to "split".