Canvas Dimensions update
Previously, canvas width and height attributes were directly set in the html tag of the canvas itself. After researching this, it was found that these attributes control the resolution of the canvas area, rather than its actual size, which is determined in the css of the element, rather than its HTML.
These attribute definitions have now been removed and are now set via JavaScript to match the size of the canvas element itself on game initialisation.